Refactor pokemon models into package
