+++ tests/test_teach_move_command.py
@@ -101,7 +101,9 @@ def setup_modules():
 
     orig_pokemon = sys.modules.get("pokemon")
     orig_gen = sys.modules.get("pokemon.generation")
-    orig_models = sys.modules.get("pokemon.models")
+    orig_models_pkg = sys.modules.get("pokemon.models")
+    orig_models_moves = sys.modules.get("pokemon.models.moves")
+    orig_models_trainer = sys.modules.get("pokemon.models.trainer")
     orig_stats = sys.modules.get("pokemon.stats")
     orig_dex = sys.modules.get("pokemon.dex")
     orig_breeding = sys.modules.get("pokemon.breeding")
@@ -111,7 +113,7 @@ def setup_modules():
     gen_mod = types.ModuleType("pokemon.generation")
     gen_mod.generate_pokemon = lambda *a, **k: None
     gen_mod.get_valid_moves = lambda species, lvl: ["tackle", "growl"]
-    models_mod = types.ModuleType("pokemon.models")
+    models_pkg = types.ModuleType("pokemon.models")
 
     class FakeMove:
         def __init__(self, name):
@@ -121,8 +123,10 @@ def setup_modules():
         def get_or_create(self_inner, name):
             return FakeMove(name), True
 
-    models_mod.Move = type("Move", (), {"objects": Manager()})
-    models_mod.InventoryEntry = type("InventoryEntry", (), {})
+    moves_mod = types.ModuleType("pokemon.models.moves")
+    moves_mod.Move = type("Move", (), {"objects": Manager()})
+    trainer_mod = types.ModuleType("pokemon.models.trainer")
+    trainer_mod.InventoryEntry = type("InventoryEntry", (), {})
     stats_mod = types.ModuleType("pokemon.stats")
     stats_mod.calculate_stats = lambda *a, **k: {}
     dex_mod = types.ModuleType("pokemon.dex")
@@ -130,14 +134,16 @@ def setup_modules():
     breeding_mod = types.ModuleType("pokemon.breeding")
 
     pokemon_pkg.generation = gen_mod
-    pokemon_pkg.models = models_mod
+    pokemon_pkg.models = models_pkg
     pokemon_pkg.stats = stats_mod
     pokemon_pkg.dex = dex_mod
     pokemon_pkg.breeding = breeding_mod
 
     sys.modules["pokemon"] = pokemon_pkg
     sys.modules["pokemon.generation"] = gen_mod
-    sys.modules["pokemon.models"] = models_mod
+    sys.modules["pokemon.models"] = models_pkg
+    sys.modules["pokemon.models.moves"] = moves_mod
+    sys.modules["pokemon.models.trainer"] = trainer_mod
     sys.modules["pokemon.stats"] = stats_mod
     sys.modules["pokemon.dex"] = dex_mod
     sys.modules["pokemon.breeding"] = breeding_mod
@@ -168,7 +174,9 @@ def setup_modules():
         orig_evennia,
         orig_pokemon,
         orig_gen,
-        orig_models,
+        orig_models_pkg,
+        orig_models_moves,
+        orig_models_trainer,
         orig_stats,
         orig_dex,
         orig_breeding,
@@ -177,7 +185,19 @@ def setup_modules():
     )
 
 
-def restore_modules(orig_evennia, orig_pokemon, orig_gen, orig_models, orig_stats, orig_dex, orig_breeding, orig_inv, orig_learn):
+def restore_modules(
+    orig_evennia,
+    orig_pokemon,
+    orig_gen,
+    orig_models_pkg,
+    orig_models_moves,
+    orig_models_trainer,
+    orig_stats,
+    orig_dex,
+    orig_breeding,
+    orig_inv,
+    orig_learn,
+):
     if orig_evennia is not None:
         sys.modules["evennia"] = orig_evennia
     else:
@@ -194,10 +214,18 @@ def restore_modules(orig_evennia, orig_pokemon, orig_gen, orig_models, orig_stat
         sys.modules["pokemon.generation"] = orig_gen
     else:
         sys.modules.pop("pokemon.generation", None)
-    if orig_models is not None:
-        sys.modules["pokemon.models"] = orig_models
+    if orig_models_pkg is not None:
+        sys.modules["pokemon.models"] = orig_models_pkg
     else:
         sys.modules.pop("pokemon.models", None)
+    if orig_models_moves is not None:
+        sys.modules["pokemon.models.moves"] = orig_models_moves
+    else:
+        sys.modules.pop("pokemon.models.moves", None)
+    if orig_models_trainer is not None:
+        sys.modules["pokemon.models.trainer"] = orig_models_trainer
+    else:
+        sys.modules.pop("pokemon.models.trainer", None)
     if orig_stats is not None:
         sys.modules["pokemon.stats"] = orig_stats
     else:
